+++ tests/smoke/property-info.html
@@ -85,7 +85,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 
   function logInfo(name, info, check1, check2) {
     var out = ['<tr>'];
-    //out.push('<td>' + name + '</td>');
+    out.push('<td>' + name + '</td>');
     out.push('<td>');
     for (var p in info) {
       var desc = info[p];
